vidyamitra-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.0
email-validator==2.3.0
pydantic-settings==2.0.0
nltk==3.8.1
python-dotenv==1.0.0
requests==2.31.0
PyMuPDF==1.24.10
python-docx==1.1.0
openai>=1.13.3
groq>=0.18.0
supabase>=2.3.0
mediapipe==0.10.14
opencv-python-headless==4.11.0.86